Refactor wizard into sections
The metadata section and wizard container have been refactored out of
the InfoBarWizard into the Wizard itself, with the InfoBarWizard now
only containing the InfoBarWizard-specific fields.

This will allow us to integrate wizards for each message type, instead
of repeating fields everywhere.
