+++ src/components/Wizard/index.tsx
@@ -7,11 +7,20 @@
 import { useState } from "react";
 import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
+import { useForm, FormProvider } from "react-hook-form";
+import Card from "react-bootstrap/Card";
+import CloseButton from "react-bootstrap/CloseButton";
+import Container from "react-bootstrap/Container";
+import Form from "react-bootstrap/Form";
+import ListGroup from "react-bootstrap/ListGroup";
 import JSONPretty from "react-json-pretty";
 
 import NewEditImportPane from "./NewEditImportPane";
 import { MessageTemplate } from "./messageTypes";
 import InfoBarWizard from "./InfoBarWizard";
+import WizardFormData from "./formData";
+import WizardMetaSection from "./WizardMetaSection";
+import serializeMessage from "./serializers";
 
 type MessageInfo = {
   id: string;
@@ -56,6 +65,7 @@ export default function Wizard() {
     undefined
   );
   const [previewJson, setPreviewJson] = useState<object | undefined>(undefined);
+  const formContext = useForm<WizardFormData>();
 
   if (typeof messageInfo === "undefined") {
     const handleNewMessage = (id: string, template: MessageTemplate) =>
@@ -67,10 +77,10 @@ export default function Wizard() {
     return <NewEditImportPane onNewMessage={handleNewMessage} />;
   }
 
-  let MessageWizard;
+  let MessageContentWizard;
   switch (messageInfo.template) {
     case "infobar":
-      MessageWizard = InfoBarWizard;
+      MessageContentWizard = InfoBarWizard;
       break;
 
     case "cfr":
@@ -82,13 +92,92 @@ export default function Wizard() {
   const stopEditing = () => setMessageInfo(undefined);
   const closeModal = () => setPreviewJson(undefined);
 
+  const { trigger, getValues } = formContext;
+
+  const validate = async (): Promise<object | null> => {
+    if (await trigger(undefined, { shouldFocus: true })) {
+      return serializeMessage(
+        messageInfo.id,
+        messageInfo.template,
+        getValues()
+      );
+    }
+
+    return null;
+  };
+
+  const handleShowJson = async (): Promise<void> => {
+    try {
+      const json = await validate();
+      if (json) {
+        setPreviewJson(json);
+      }
+    } catch (e) {
+      console.error(e);
+    }
+  };
+
+  const handlePreview = async (): Promise<void> => {
+    try {
+      const json = await validate();
+      if (json) {
+        const jsonStr = JSON.stringify(json);
+        const uri = `about:messagepreview?json=${btoa(jsonStr)}`;
+
+        return navigator.clipboard.writeText(uri);
+      }
+    } catch (e) {
+      console.error(e);
+    }
+  };
+
   return (
     <>
-      <MessageWizard
-        id={messageInfo.id}
-        stopEditing={stopEditing}
-        setPreviewJson={setPreviewJson}
-      />
+      <Container className="wizard">
+        <FormProvider {...formContext}>
+          <Form>
+            <Card>
+              <Card.Header className="d-flex justify-content-between">
+                <Card.Title className="mb-0">
+                  Editing Message:{" "}
+                  <span className="message-id">{messageInfo.id}</span>
+                </Card.Title>
+                <CloseButton onClick={stopEditing} title="Stop Editing" />
+              </Card.Header>
+
+              <ListGroup variant="flush">
+                <ListGroup.Item className="wizard-section-header">
+                  Message Content
+                </ListGroup.Item>
+
+                <ListGroup.Item>
+                  <MessageContentWizard />
+                </ListGroup.Item>
+
+                <ListGroup.Item className="wizard-section-header">
+                  Metadata
+                </ListGroup.Item>
+
+                <ListGroup.Item>
+                  <WizardMetaSection />
+                </ListGroup.Item>
+
+                <ListGroup.Item className="wizard-buttons">
+                  <Button onClick={() => void handleShowJson()}>
+                    Show JSON
+                  </Button>
+                  <Button
+                    onClick={() => void handlePreview()}
+                    className="copy-button"
+                  >
+                    Copy Preview Link
+                  </Button>
+                </ListGroup.Item>
+              </ListGroup>
+            </Card>
+          </Form>
+        </FormProvider>
+      </Container>
       <JsonPreview data={previewJson} onHide={closeModal} />
     </>
   );
