+++ src/components/Wizard/messageTypes.ts
@@ -4,8 +4,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import LocalizableTextFormData from "../LocalizableTextInput/formData";
-
 export type MessageTemplate =
   | "cfr"
   | "infobar"
@@ -13,9 +11,21 @@ export type MessageTemplate =
   | "pbnewtab";
 
 export type LocalizableText = string | { string_id: string };
-export interface InfoBarMessage {
+export interface InfoBarMessageContent {
+  type: "tab" | "global";
+  text: LocalizableText;
+  priority?: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
+  buttons: {
+    label: LocalizableText;
+    primary?: boolean;
+    accessKey?: string;
+    supportPage?: string;
+    action: object;
+  }[];
+}
+export interface BaseMessage {
   id: string;
-  template: "infobar";
+  template: MessageTemplate;
   targeting: string;
   groups?: string[];
   trigger: object;
@@ -28,26 +38,9 @@ export interface InfoBarMessage {
       cap: number;
     }[];
   };
-  content: {
-    type: "tab" | "global";
-    text: LocalizableText;
-    priority?: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
-    buttons: {
-      label: LocalizableText;
-      primary?: boolean;
-      accessKey?: string;
-      supportPage?: string;
-      action: object;
-    }[];
-  };
 }
 
-export function localizableTextToJson(
-  data: LocalizableTextFormData
-): LocalizableText {
-  if (data.localized) {
-    return { string_id: data.stringId };
-  } else {
-    return data.text;
-  }
-}
+export type Message = BaseMessage & {
+  template: "infobar";
+  content: InfoBarMessageContent;
+};
